refactor(mcp): split mcp into submodules (M3-REF-003)
